delete user and videos and comments and add video to a playlist

+++ Aparat/app/Jobs/ConvertAndAddWaterMarkToUploadedVideoJob.php
@@ -11,7 +11,9 @@ use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Http\Request;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Queue\SerializesModels;
+use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Storage;
+use Lcobucci\JWT\Exception;
 use ProtoneMedia\LaravelFFMpeg\Filesystem\Media;
 
 class ConvertAndAddWaterMarkToUploadedVideoJob implements ShouldQueue
@@ -42,33 +44,39 @@ class ConvertAndAddWaterMarkToUploadedVideoJob implements ShouldQueue
     public function handle(): void
     {
 
+            $upladedVideoPath = '/tmp/' . $this->videoId;
 
-        $upladedVideoPath='/tmp/' . $this->videoId;
-        $videoUploaded = \FFM::fromDisk('videos')->open($upladedVideoPath);
+            if ($this->video->trashed() || !Video::where('id',$this->videoId)->count()) {
+                Storage::disk('videos')->delete($upladedVideoPath);
+                return;
+            }
 
+            $videoUploaded = \FFM::fromDisk('videos')->open($upladedVideoPath);
+            $format = new \FFMpeg\Format\Video\X264('libmp3lame');
 
-        $format = new \FFMpeg\Format\Video\X264('libmp3lame');
+            /** @var Media $videoFile */
 
-        /** @var Media $videoFile */
-
-        if ($this->addWatermark){
-            $filter=new CustomFilter("drawtext=text='http\\://aliattar.com'
+            if ($this->addWatermark) {
+                $filter = new CustomFilter("drawtext=text='http\\://aliattar.com'
              :fontcolor=white:fontsize=30:box=1:boxcolor=white@0.5");
 
-            $videoUploaded = $videoUploaded->addFilter($filter);
+                $videoUploaded = $videoUploaded->addFilter($filter);
+            }
+
+            /** @var Media $videoFile */
+            $videoFile = $videoUploaded->export()
+                ->toDisk('videos')
+                ->inFormat($format);
+
+            $videoFile->save($this->userId . '/' . $this->video->slug . '.mp4');
+
+            $this->video->duration = $videoUploaded->getDurationInSeconds();
+            $this->video->state = Video::STATE_CONVERTED;
+            $this->video->save();
+
+            Storage::disk('videos')->delete($upladedVideoPath);
+
+
+
         }
-
-        /** @var Media $videoFile */
-        $videoFile = $videoUploaded->export()
-            ->toDisk('videos')
-            ->inFormat($format);
-
-        $videoFile->save($this->userId . '/' . $this->video->slug . '.mp4');
-
-        $this->video->duration =$videoUploaded->getDurationInSeconds();
-        $this->video->state = Video::STATE_CONVERTED;
-        $this->video->save();
-
-        Storage::disk('videos')->delete($upladedVideoPath);
-    }
 }
